Common WAN helper functions
