+++ diffusers_nodes_library/common/parameters/diffusion/wan/img2vid_runtime_parameters.py
@@ -1,22 +1,15 @@
 import logging
-import tempfile
-from pathlib import Path
 from typing import Any
 
-import diffusers  # type: ignore[reportMissingImports]
 import numpy as np
-import PIL.Image
-import torch  # type: ignore[reportMissingImports]
 from griptape.artifacts import ImageUrlArtifact
-from griptape.artifacts.video_url_artifact import VideoUrlArtifact
 from griptape.loaders import ImageLoader
 from griptape_nodes.exe_types.core_types import Parameter, ParameterMode
 from griptape_nodes.exe_types.node_types import BaseNode
-from griptape_nodes.files.project_file import ProjectFileDestination
 from PIL.Image import Image
 
-from diffusers_nodes_library.common.parameters.diffusion.runtime_parameters import (
-    DiffusionPipelineRuntimeParameters,
+from diffusers_nodes_library.common.parameters.diffusion.wan.base_runtime_parameters import (
+    WanVideoPipelineRuntimeParametersBase,
 )
 from pillow_nodes_library.utils import (  # type: ignore[reportMissingImports]
     image_artifact_to_pil,
@@ -25,7 +18,7 @@ from pillow_nodes_library.utils import (  # type: ignore[reportMissingImports]
 logger = logging.getLogger("diffusers_nodes_library")
 
 
-class WanImageToVideoPipelineRuntimeParameters(DiffusionPipelineRuntimeParameters):
+class WanImageToVideoPipelineRuntimeParameters(WanVideoPipelineRuntimeParametersBase):
     def __init__(self, node: BaseNode):
         super().__init__(node)
 
@@ -123,21 +116,12 @@ class WanImageToVideoPipelineRuntimeParameters(DiffusionPipelineRuntimeParameter
             "guidance_scale": self._node.get_parameter_value("guidance_scale"),
         }
 
-    def publish_output_image_preview_placeholder(self) -> None:
-        # Create a small black video placeholder
-        with tempfile.NamedTemporaryFile(suffix=".mp4", delete=False) as temp_file:
-            temp_path = Path(temp_file.name)
-        try:
-            # Create a single black frame and export as 1-frame video
-            black_frame = PIL.Image.new("RGB", (320, 240), color="black")
-            frames = [black_frame]
-            diffusers.utils.export_to_video(frames, str(temp_path), fps=1)
-            dest = ProjectFileDestination.from_situation(filename="placeholder_video.mp4", situation="save_node_output")
-            saved = dest.write_bytes(temp_path.read_bytes())
-            self._node.publish_update_to_parameter("output_video", VideoUrlArtifact(saved.location))
-        finally:
-            if temp_path.exists():
-                temp_path.unlink()
+    def _get_pipeline_call_kwargs(self, pipe: Any) -> dict:
+        """Override to include image processing for image-to-video pipeline."""
+        from diffusers_nodes_library.common.utils.pipeline_utils import get_repo_id_from_pipeline
+
+        repo_id = get_repo_id_from_pipeline(pipe)
+        return self.get_pipe_kwargs_with_image(pipe, repo_id)
 
     def get_input_image_pil(self) -> Image:
         input_image_artifact = self._node.get_parameter_value("input_image")
@@ -185,7 +169,7 @@ class WanImageToVideoPipelineRuntimeParameters(DiffusionPipelineRuntimeParameter
 
         return image, height, width
 
-    def get_pipe_kwargs(self, pipe: Any, repo_id: str) -> dict:
+    def get_pipe_kwargs_with_image(self, pipe: Any, repo_id: str) -> dict:
         image, height, width = self.get_image_for_model(pipe, repo_id)
         return {
             **self._get_pipe_kwargs(),
@@ -193,59 +177,3 @@ class WanImageToVideoPipelineRuntimeParameters(DiffusionPipelineRuntimeParameter
             "height": height,
             "width": width,
         }
-
-    def latents_to_video_mp4(self, pipe: Any, latents: Any) -> Path:
-        """Convert latents to video frames and export as MP4 file."""
-        with tempfile.NamedTemporaryFile(suffix=".mp4", delete=False) as temp_file_obj:
-            temp_file = Path(temp_file_obj.name)
-
-        try:
-            # Convert latents to video frames using VAE decode
-            latents = latents.to(pipe.vae.dtype)
-
-            # Apply latents normalization as per the WAN pipeline
-            latents_mean = (
-                torch.tensor(pipe.vae.config.latents_mean)
-                .view(1, pipe.vae.config.z_dim, 1, 1, 1)
-                .to(latents.device, latents.dtype)
-            )
-            latents_std = 1.0 / torch.tensor(pipe.vae.config.latents_std).view(1, pipe.vae.config.z_dim, 1, 1, 1).to(
-                latents.device, latents.dtype
-            )
-            latents = latents / latents_std + latents_mean
-
-            # Decode latents to video using VAE
-            video = pipe.vae.decode(latents, return_dict=False)[0]
-            frames = pipe.video_processor.postprocess_video(video, output_type="pil")[0]
-
-            # Export frames to video
-            diffusers.utils.export_to_video(frames, str(temp_file), fps=16)
-        except Exception:
-            # Clean up on error
-            if temp_file.exists():
-                temp_file.unlink()
-            raise
-        else:
-            return temp_file
-
-    def publish_output_video_preview_latents(self, pipe: Any, latents: Any) -> None:
-        """Publish a preview video from latents during generation."""
-        preview_video_path = None
-        try:
-            preview_video_path = self.latents_to_video_mp4(pipe, latents)
-            dest = ProjectFileDestination.from_situation(filename="preview_video.mp4", situation="save_node_output")
-            saved = dest.write_bytes(preview_video_path.read_bytes())
-            self._node.publish_update_to_parameter("output_video", VideoUrlArtifact(saved.location))
-        except Exception as e:
-            logger.warning("Failed to generate video preview from latents: %s", e)
-        finally:
-            # Clean up temporary file
-            if preview_video_path is not None and preview_video_path.exists():
-                preview_video_path.unlink()
-
-    def publish_output_video(self, video_path: Path) -> None:
-        dest = ProjectFileDestination.from_situation(
-            filename=f"output_video{video_path.suffix}", situation="save_node_output"
-        )
-        saved = dest.write_bytes(video_path.read_bytes())
-        self._node.parameter_output_values["output_video"] = VideoUrlArtifact(saved.location)
